+++ packages/auto_animated/example/lib/utils.dart
@@ -38,7 +38,7 @@ class HorizontalItem extends StatelessWidget {
             child: Center(
               child: Text(
                 title,
-                style: Theme.of(context).textTheme.headline4,
+                style: Theme.of(context).textTheme.headlineMedium,
               ),
             ),
           ),