+++ packages/epub_view/lib/src/helpers/epub_document.dart
@@ -1,5 +1,6 @@
 // ignore: unnecessary_import
 import 'dart:typed_data';
+import 'package:http/http.dart' as http;
 
 import 'package:epubx/epubx.dart';
 import 'package:flutter/services.dart';
@@ -19,5 +20,16 @@ class EpubDocument {
   static Future<EpubBook> openFile(File file) async {
     final bytes = await file.readAsBytes();
     return EpubReader.readBook(bytes);
+  }
+
+  static Future<EpubBook> openUrl(String url) async {
+    var uri = Uri.parse(url);
+    final response = await http.get(uri);
+
+    if (response.statusCode == 200) {
+      return EpubReader.readBook(response.bodyBytes);
+    } else {
+      throw Exception('Failed to load epub');
+    }
   }
 }