+++ packages/pdfx/darwin/pdfx/Sources/pdfx/Hooks.swift
@@ -1,4 +1,7 @@
+import CoreGraphics
+import Foundation
 #if os(iOS)
+import UIKit
 
 extension UIColor {
     convenience init(hexString: String) {
@@ -21,6 +24,7 @@ extension UIColor {
 }
 
 #elseif os(macOS)
+import AppKit
 
 extension NSColor {
     convenience init(hexString: String) {