+++ packages/pdfx/ios/Classes/PdfxPlugin.h
@@ -1,8 +1,0 @@
-#if TARGET_OS_IOS
-#import <Flutter/Flutter.h>
-#else
-#import <FlutterMacOS/FlutterMacOS.h>
-#endif
-
-@interface PdfxPlugin : NSObject <FlutterPlugin>
-@end