+++ packages/pdfx/lib/src/renderer/web/pdfjs.dart
@@ -16,9 +16,6 @@
 
 // ignore_for_file: public_member_api_docs
 
-@JS()
-library pdf.js;
-
 import 'dart:js_interop';
 import 'dart:typed_data';
 
@@ -34,7 +31,8 @@ external PdfjsLib? get _pdfjsLib;
 external PdfjsRenderOptions? get _pdfRenderOptionsFromContext;
 
 // Safe check
-PdfjsRenderOptions get _pdfRenderOptions => _pdfRenderOptionsFromContext ?? Constants.defaultPdfjsRenderOptions;
+PdfjsRenderOptions get _pdfRenderOptions =>
+    _pdfRenderOptionsFromContext ?? Constants.defaultPdfjsRenderOptions;
 
 extension type PdfjsLib(JSObject _) implements JSObject {
   external PdfjsDocumentTask getDocument(PdfjsRenderOptions options);
@@ -135,9 +133,9 @@ extension type PdfjsViewportParams._(JSObject _) implements JSObject {
   external factory PdfjsViewportParams({
     double scale,
     int rotation, // 0, 90, 180, 270
-    double offsetX = 0,
-    double offsetY = 0,
-    bool dontFlip = false,
+    double offsetX,
+    double offsetY,
+    bool dontFlip,
   });
 
   external double get scale;
@@ -204,9 +202,9 @@ extension type PdfjsRenderContext._(JSObject _) implements JSObject {
   external factory PdfjsRenderContext({
     required web.CanvasRenderingContext2D canvasContext,
     required PdfjsViewport viewport,
-    String intent = 'display',
-    bool enableWebGL = false,
-    bool renderInteractiveForms = false,
+    String intent,
+    bool enableWebGL,
+    bool renderInteractiveForms,
     JSArray<JSNumber>? transform,
     JSAny imageLayer,
     JSAny canvasFactory,