+++ packages/pdfx/macos/Classes/SwiftPdfxPlugin.swift
@@ -1,1 +1,0 @@
-../../ios/Classes/SwiftPdfxPlugin.swift