+++ packages/pdfx/macos/Classes/pdfx.h
@@ -1,1 +1,0 @@
-../../ios/Classes/pdfx.h