+++ packages/epub_view/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 3.3.0
+
+* Added support for opening books from a URL via `EpubDocument.openUrl` [pull#470](https://github.com/ScerIO/packages.flutter/pull/470)
+
 ## 3.2.0
 
 * Upgrade epubx dependency